Update to test code
Filled out the code for writing and reading to various devices as well as on-start initialization. Working on configuration of FT2232HL and rewriting sample code to avoid potential licensing concerns. A few functions named in the code are pulled from the libMPSSE sample file and I'd like to get them out of there so this is all my original code. No license for the code was provided except for a copyright notice at the top of the sample code, so to avoid the issue altogether I plan on rewriting read and write functions, and getting rid of the timing functions.